better way to deal with net::ERR_NETWORK_CHANGED
Before, we relied on a total hack to tell if the error happened or not (checking for window.muniverse). On my Mac, this worked. On my Linux machine, however, sometimes net::ERR_NETWORK_CHANGED allowed the main script to load but caused problems loading things with AJAX (the log I saw was only complaining about 'combo.ogg'). This is probably a timing issue. Docker is probably switching something about the network stack *after* loading the OS, and that sometimes messes up Chrome's page loading.