better container garbage collection
Now if a the program dies between the time of container creation and connecting to the kill socket, the container will still die after a timeout. Since netcat that I use didn't have a listen timeout feature, I had to roll my own netcat-like tool (it's a short Go program).