give kill socket listener more time to startup
Technically there is a race condition where the container starts and we try to establish the kill socket before the listener is running. In this case, the container will die after the kill socket listen timeout. I have never experienced this case, but this change makes the case virtually impossible (although not theoretically impossible). Ideally, there would be a reconnect system that could wait for the kill socket to actually start running. The challenge is that Docker will accept our connection, even if the kill socket server is not running yet.