synchronize docker calls
This is a workaround for a Docker race condition in which two containers are assigned the same IP address. In that case, two chrome clients would attempt to connect to the same headless chrome instance, resulting in one of them failing to connect (the API only allows one client at a time). By synchronizing docker calls, we can be more certain that the race condition will not occur, although it may still occur with multiple muniverse instances running in parallel.

Docker issue is here:
https://github.com/docker/libnetwork/issues/1740