game variant system
This is necessary to stay true to some games, which gave different levels or pre-game options. It can also be thought of as a form of data augmentation.

The design of the variant system was carefully thought out; it is the result of a lot of tradeoffs. I had the following goals in mind:

 * Backwards-compatibility with older environments & muniverse code
 * Don't require re-downloads of the same game
 * Make it easy (minimal typing) to define variants, especially ones which only differ in screen resolution.

Each of these goals is achieved:

 * The EnvSpec structure was extended but the old fields are unchanged.
 * Variants reference a "base game" and use its resources.
 * All attributes are inherited from base game by default.