added Shards-v0
This game was a total pain to integrate. For one thing, it lazily loads background images for each level. It also totally lags out my machine. Add to all this that, for some reason, it doesn't allow headless Chromium to play the game in portrait mode (and thus required some patching). Plus, the game is like 10MB.