drafted protocol for language bindings
The binding system, as I am building it, will be based on command execution. A command, called "muniverse-bind", will be run by another language in order to control muniverse environments. To see why I chose this model, here were my alternatives:

 1. Dynamically link Go code to other languages using CGO.
 2. Create a server/client API, like gym-http-api.
 3. Rewrite muniverse in Python.

I did not like option 3 because Python is slow when it comes to driving concurrent environments. OpenAI discovered this when building Universe, and ended up writing a VNC client in Go to compensate. I'd like to avoid bottlenecks associated with the Python GIL. Also, and this goes without saying, other languages would not be supported.

I don't like option 2 because a) it requires the user to run a separate program whenever they use muniverse, b) it requires TCP port number coordination (although a default port can help with this). In practice, I have found it annoying that I have to run gym-socket-api whenever I want to use OpenAI Gym from Go. The command-based solution can be seen as a form of option 2, but with the process execution hidden from view.

Option 1 is fairly promising, but it has some drawbacks. For one, it would likely involve creating a new C/C++ shim for every new language. It would also make bindings depend very closely on the underlying Go code. With the command-based solution, the muniverse-bind command could be written in any language as long as it exposes the correct API; it can be swapped out simply by changing the $PATH. Using dynamic linking would also allow some amount of abstraction, but not as much.