added 100BallsOnline-v0
This was an interesting game because it had an unobfuscated version of Construct2 version 1.0. It was really neat to see what's going on inside of Construct2 a little better. Also, the game hooked its own ad API into Construct2 via the C2 plugin API, which I've never had to deal with before.