benchmark observations
Capturing observations is the main bottleneck in muniverse environments. The benchmarks show that PNG observations are nearly twice as slow as JPEG observations, and even then both types of observations are extremely slow. The overhead is likely in Chrome, either from encoding the image as raw data or as encoding the raw data as base64+JSON+websocket.

I will use this benchmark as I experiment with other versions of Chrome, as well as the startScreencast API. I am not sure if I'll be able to use the screencast API due to its async nature, but perhaps it will come in handy.