don't use external font in DontCrash-v0
No environment should require the internet. As I was running tests, I noticed that DontCrash-v0 was connecting to Google's font server. This patch pre-downloads the font to avoid that.

In general, I try to be careful about checking that games don't connect to the internet or phone home. Back when I implemented DontCrash-v0 (one of my first games), I didn't know I had to be as strict about this sort of thing.