faster resets after first reset
This shaves almost exactly a second off of every reset (except the first). For many environments, this cuts the reset time nearly in half.

Before, I was using NavigateSafe for every reset. Now, I only use it on the first reset, which is the only time it should ever be needed (to deal with initial Docker networking issues). The NavigateSafe method waits an extra second compared to NavigateSync (as part of being "safe"). Thus, by not using it, we save a second for free.